Another restructure and refactor
