+++ FlutterApp/lib/resources/classify_screen/result_card.dart
