+++ FlutterApp/lib/resources/classify_screen/sub_screens/load_image.dart
@@ -5,8 +5,9 @@ import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:new_ui/resources/common/load_image_dialog.dart';
 
-import '../../../components/button.dart';
-import '../functions.dart';
+import '../../common/button.dart';
+import '../../common/page_indicator.dart';
+import '../methods.dart';
 
 class LoadImageScreen extends StatefulWidget {
   final PageController controller;
@@ -26,7 +27,7 @@ class _LoadImageScreen extends State<LoadImageScreen> {
     setState(() {
       if (kDebugMode) {
         print("rebuild!");
-        print(LoadImageScreen.pickedImage);
+        //print(LoadImageScreen.pickedImage);
       }
     });
   }
@@ -53,6 +54,7 @@ class _LoadImageScreen extends State<LoadImageScreen> {
         padding: const EdgeInsets.only(
             left: 13.0, right: 13.0, bottom: 13.0, top: 45),
         children: [
+          classifyPageIndicator(controller: widget.controller, count: 2),
           const SizedBox(
             height: 30,
           ),
