+++ FlutterApp/lib/resources/common/globals.dart
