+++ FlutterApp/lib/resources/common/nav_bar.dart
@@ -1,0 +1,1 @@
+// TODO IMPLEMENT NAVBAR WIDGET HERE
