+++ FlutterApp/lib/resources/common/suggestions.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 
 import 'package:http/http.dart' as http;
 
-import '../functions/func.dart';
+import 'func.dart';
 
 String domain = getDomain(1);
