+++ FlutterApp/lib/screens/add.dart
@@ -9,10 +9,10 @@ import 'package:flutter/services.dart';
 import 'package:http/http.dart' as http;
 import 'package:http/io_client.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:new_ui/components/button.dart';
-import 'package:new_ui/components/globals.dart' as globals;
-import 'package:new_ui/functions/func.dart';
-import 'package:new_ui/screens/mealsuggestions.dart';
+import 'package:new_ui/resources/common/button.dart';
+import 'package:new_ui/resources/common/func.dart';
+import 'package:new_ui/resources/common/globals.dart' as globals;
+import 'package:new_ui/screens/suggestions.dart';
 import 'package:universal_platform/universal_platform.dart';
 
 class AddImage extends StatefulWidget {
