+++ FlutterApp/lib/screens/home.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
-import '../components/button.dart';
+import '../resources/common/button.dart';
 
 class Home extends StatelessWidget {
   const Home({
